fix: combine minify.js, worker.js for ncc interop
